+++ keel-engine-epoll/src/linuxMain/kotlin/io/github/fukusaka/keel/engine/epoll/EpollEventLoop.kt
@@ -31,7 +31,6 @@ import kotlinx.cinterop.ptr
 import kotlinx.cinterop.staticCFunction
 import kotlinx.coroutines.CancellableContinuation
 import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.suspendCancellableCoroutine
 import platform.linux.EPOLLERR
 import platform.linux.EPOLLHUP
 import platform.linux.EPOLLIN
@@ -657,15 +656,7 @@ internal class EpollEventLoop(
 
     // --- EpollSuspendRegister impl (seam for connect InProgress) ---
 
-    override suspend fun awaitWriteReady(fd: Int, logger: Logger) {
-        suspendCancellableCoroutine<Unit> { cont ->
-            val reg = register(fd, Interest.WRITE, cont)
-            cont.invokeOnCancellation {
-                unregister(reg)
-                closeFdSafely(fd, logger, "connect cancellation")
-            }
-        }
-    }
+    override suspend fun awaitWriteReady(fd: Int, logger: Logger) = awaitWritableOwningFd(fd, logger)
 
     companion object {